+++ examples/example3.cpp
@@ -56,7 +56,7 @@ int main(int argc, char *argv[]) {
         std::cout << "#############################" << std::endl;
         std::cout << cameraWrapper.getSummary() << std::endl;
     } catch (gphoto2pp::exceptions::gphoto2_exception &e) {
-        std::cout << "GPhoto Exception Code: " << e.getResultCode() << std::endl;
+        std::cout << "GPhoto Exception Code: " << static_cast<int>(e.getResultCode()) << std::endl;
         std::cout << "Exception Message: " << e.what() << std::endl;
     }
 }
